clean up warnings (#1269)
* warnings in sdo, moxy, dbws
* add override ann
* remove redundant casts, inheritdoc

+++ foundation/org.eclipse.persistence.core/src/main/java/org/eclipse/persistence/internal/expressions/ExpressionJavaPrinter.java
@@ -144,11 +144,11 @@ public class ExpressionJavaPrinter {
             printString("}");
         } else if (object.getClass() == ClassConstants.BIGDECIMAL) {
             printString("new java.math.BigDecimal(\"");
-            printString(((BigDecimal)object).toString());
+            printString(object.toString());
             printString("\")");
         } else if (object.getClass() == ClassConstants.BIGINTEGER) {
             printString("new java.math.BigInteger(\"");
-            printString(((BigInteger)object).toString());
+            printString(object.toString());
             printString("\")");
         } else {
             printString((String)ConversionManager.getDefaultManager().convertObject(object, String.class));
